refactor(cli): move process-exit lifecycle to CLI boundary
